xkbcomp: Ensure XKB canonical key types presence
The following types are called “Canonical Key Types” and the XKB protocol
specifies them as mandatory in any keymap:

- ONE_LEVEL
- TWO_LEVEL
- ALPHABETIC
- KEYPAD

They must have specific properties defined in the appendix B of
“The X Keyboard Extension: Protocol Specification”:
https://www.x.org/releases/current/doc/kbproto/xkbproto.html#canonical_key_types

In the Xorg ecosystem, any missing canonical type fallbacks to a default
type supplied by libX11’s `XkbInitCanonicalKeyTypes()`, e.g. in xkbcomp.

libxkbcommon does not require these types per se: it only requires that
all *used* types — explicit (`type="…"`) or implicit (automatic types) —
are defined, with the exception that if no key type at all is defined,
then a default one-level type is provided.

libxkbcommon also does not require any particular order of these key
types, because they are retrieved using their name instead of their index.

While the canonical key types are always present in the keymap generated
from xkeyboard-config, since 31900860c65b88e4d10ad7dd00377e2815cca0f6 we
drop unused key types by default, which may happen with some layouts
with 4+ levels may not require e.g. the `TWO_LEVEL` nor the `ALPHABETIC`
types.

In theory, libxkbcommon would not care of the presence of the canonical
key types and could delegate the fallback and ordering work to xkbcomp,
as it is the case in Xorg’s Xwayland. However the implementation is buggy:

- https://gitlab.freedesktop.org/xorg/lib/libx11/-/merge_requests/292
- https://gitlab.freedesktop.org/xorg/xserver/-/merge_requests/2082

It prevents the keymap to be properly uploaded in the server, leading to
errors when retrieving it with libxkbcommon-x11. See:
https://gitlab.archlinux.org/archlinux/packaging/packages/libxkbcommon/-/issues/3

Since it’s not clear when new versions of libX11 and xserver will be
released, this commit ensures the presence of the canonical key types.

WARNING: A keymap that uses them (explicitly `type="…"` or implicitly
with automatic types) without providing them is considered buggy.