context: Add default include paths only if necessary
Default includes paths are delayed and added only if necessary.

It is more efficient for most clients, which only get the keymap from
the server: it avoids unnecessary allocations and file system queries.

It also avoid the corner case where a containerized app does not have
access to the XKB directories.

There might be an issue in case the environment variables relevant to
the include paths change between the context initialization and the
call to `xkb_context_include_path_append_default()`. However this seems
very unlikely and would have already triggered issues when the `%`-
expansion is used.