context: Add legacy X11 include path for misconfigured setups
Some setups use the assumption that the canonical XKB root is always
the legacy X11 one, but this is no longer true since xkeyboard-config
2.45, where the X11 path is now a mere symlink to a dedicated data
directory for xkeyboard-config.

Fixed by using a fallback to the legacy X11 path for misconfigured
setups where the canonical XKB root is not available.

This fallback can still be skipped if the environment variable
`XKB_CONFIG_ROOT` is deliberately set to an empty string.