keymap: Rename long key/alias names in keymap format v1
Rename keys and aliases with names longer than 4 characters when
serializing using `XKB_KEYMAP_FORMAT_TEXT_V1`, to ensure compatibility
with the X11 ecosystem.

The new name is one of the following 2 forms:
- If key name and keycode ≤ 0xfff: prefix character + hexadecimal keycode
- If alias or keycode > 0xfff: prefix character + hexadecimal incrementing
  counter.

Prefixes are chosen in the set “!#$%&'()*+,-./0” and avoid any name
conflict.

In the very unlikely case that renaming fails due to unavailable prefixes,
the original name is kept so that Wayland clients can still parse the
keymap.