rules: Add new lookup mechanism: pre & post
Before this commit, the only way to extend rules was to override the
default rules files by creating a file with the same name in an include
directory with a higher priority than the default rules file. However,
this prevented to *compose* any rules extension by other mean than
explicit includes/inlining all rules in the higher priority rules files.

This commit introduces a new lookup mechanism that enables to extend
rules without altering the main rules file and facilitate *rules
composition*. The main motivation is to support composing custom rules
with the upcoming XKB *extensions mechanism*. This is done by enabling
*partial* rules files `<rules>.pre` and `<rules>.post`.

The previous lookup mechanism:
1. Find the rules file `<rules>`
2. Fail if not found
3. Otherwise resolves the rules
4. Fail if any required component is missing
5. Otherwise return the resolved KcCGST names

is replaced with:
1. Find the main rules file `<rules>`
2. Fail if not found
3. For each include directory:
   1. Try to find the *partial* rules file `<rules>.pre` in the directory
   2. If none is found, continue to the next directory
   3. Otherwise resolve the rules from `<rules>.pre` by appending the
      resolved components to the previous result

   This may combine *multiple* `<rules>.pre` files.
4. resolve the rules from `<rules>` by appending to the previous result;
5. same as step 3, but with the rules file `<rules>.post`.

Thus the final rule set is equivalent to:

    ! include <include path 1>/rules/<rules>.pre  // only if defined
    …
    ! include <include path n>/rules/<rules>.pre  // only if defined
    ! include <rules>                             // main rules file
    ! include <include path 1>/rules/<rules>.post // only if defined
    …
    ! include <include path n>/rules/<rules>.post // only if defined

Example: given the following tree:
- `~/.config/xkb/rules`
  - `evdev.post`
- `/etc/xkb/rules`
  - `evdev.pre`
  - `evdev.post`
- `/usr/share/xkeyboard-config-2/rules`
  - `evdev`

the new mechanism is equivalent to parsing the following rules file:

    // .pre rules
    include %E/evdev.pre // only .pre file

    // main rules
    include %S/evdev

    // .post rules
    include %H/.config/xkb/evdev.post // higher priority than /etc/xkb
    include %E/evdev.post