tools: Add --controls to interactive-evdev
This enables to configure keyboard controls. The argument value is a
comma-separated list of an optional list of controls.

Each entry can be preceeded by +/- to respectively enable or disable the
corresponding option, e.g. `--controls=+xxx,-yyy`. It ensures that
defaults can be explicitly overriden.

This is a preparatory work: no option is currently implemented.