Make use of `counted_by` gcc/clang attribute
This is kind-of nice self-documentation, and can be utilized for static
analysis and dynamic buffer overflow protection (such as
`-fsanitize=bounds` and `__builtin_dynamic_object_size` used by
`_FORTIFY_SOURCE=3`).

Clang<19.1.0 and GCC<=15.2 only support `counted_by` on flexible array
members, not pointer fields as used in xkbcommon, therefore we need a
configure check that annotating pointer fields is supported, not merely
that the attribute exists.

Clang (unlike GCC) currently requires that the count field come before
the pointer field. They have a flag to
`-fexperimental-late-parse-attributes` to enable this, but since it's
experimental I don't want to pass it. So we need to reorder some fields.

Another needed code adaption is to ensure that the length field is
always correctly set *before* the array is accessed. So e.g. the
sequence "alloc buffer, assign, set length field" is changed to "alloc
buffer, set length field, assign", otherwise the dynamic bounds check
during assignment fails.

I have verified that it works well with `-fsanitize=undefined` by
inserting some out-of-bounds access to `keymap->keys` and seeing that it
goes undetected before and detected after.

See:
https://gcc.gnu.org/onlinedocs/gcc/Common-Variable-Attributes.html#index-counted_005fby-variable-attribute
https://www.gnu.org/software/libc/manual/html_node/Source-Fortification.html#Fortification-of-function-calls
https://gcc.gnu.org/onlinedocs/gcc/Object-Size-Checking.html#index-_005f_005fbuiltin_005fdynamic_005fobject_005fsize
https://gcc.gnu.org/onlinedocs/gcc/Instrumentation-Options.html#index-fsanitize_003dbounds

Signed-off-by: Ran Benita <ran@unusedvar.com>